requirements.txt
Django==5.1.3
paho-mqtt==1.6.1
postgres
psycopg2
django-environ
gunicorn